requirements.txt
PyPDF2
keybert
openai
python-dotenv
scikit-learn 
sentence-transformers
spacy
pandas
